+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/newdeployment/DeploymentRepository.scala
@@ -1,7 +1,10 @@
 package pl.touk.nussknacker.ui.process.newdeployment
 
+import cats.implicits.{toFoldableOps, toTraverseOps}
 import db.util.DBIOActionInstances._
 import org.postgresql.util.{PSQLException, PSQLState}
+import pl.touk.nussknacker.engine.api.deployment.{DeploymentStatus, ProblemDeploymentStatus}
+import pl.touk.nussknacker.engine.newdeployment.DeploymentId
 import pl.touk.nussknacker.ui.db.entity.ProcessEntityData
 import pl.touk.nussknacker.ui.db.{DbRef, NuTables}
 import pl.touk.nussknacker.ui.process.newdeployment.DeploymentEntityFactory.DeploymentEntityData
@@ -11,10 +14,11 @@ import pl.touk.nussknacker.ui.process.newdeployment.DeploymentRepository.{
 }
 import slick.jdbc.JdbcProfile
 
-import java.sql.SQLIntegrityConstraintViolationException
+import java.sql.{SQLIntegrityConstraintViolationException, Timestamp}
+import java.time.Clock
 import scala.concurrent.ExecutionContext
 
-class DeploymentRepository(dbRef: DbRef)(implicit ec: ExecutionContext) extends NuTables {
+class DeploymentRepository(dbRef: DbRef, clock: Clock)(implicit ec: ExecutionContext) extends NuTables {
 
   override protected val profile: JdbcProfile = dbRef.profile
 
@@ -46,6 +50,28 @@ class DeploymentRepository(dbRef: DbRef)(implicit ec: ExecutionContext) extends 
         .headOption
         .map(_.map(DeploymentWithScenarioMetadata.apply _ tupled))
     )
+  }
+
+  def updateDeploymentStatuses(statusesToUpdate: Map[DeploymentId, DeploymentStatus]): DB[Set[DeploymentId]] = {
+    statusesToUpdate.toList
+      .map { case (id, status) =>
+        val problemDescription = ProblemDeploymentStatus.extractDescription(status)
+        toEffectAll(
+          deploymentsTable
+            .filter(d =>
+              d.id === id && (d.statusName =!= status.name || d.statusProblemDescription =!= problemDescription)
+            )
+            .map(d => (d.statusName, d.statusProblemDescription, d.statusModifiedAt))
+            .update((status.name, problemDescription, Timestamp.from(clock.instant())))
+            .map { result =>
+              if (result > 0) Set(id) else Set.empty[DeploymentId]
+            }
+        )
+      }
+      .sequence
+      .map(_.combineAll)
+      // For the performance reasons it is better to run all updates in the one session, transactionally should enforce it
+      .transactionally
   }
 
 }