+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/DeploymentManagerStub.scala
@@ -5,6 +5,7 @@ import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.process.{ProcessIdWithName, ProcessName}
 import pl.touk.nussknacker.engine.deployment.{DeploymentId, ExternalDeploymentId}
 import pl.touk.nussknacker.engine.management.periodic.model.PeriodicProcessDeploymentId
+import pl.touk.nussknacker.engine.newdeployment
 import pl.touk.nussknacker.engine.testing.StubbingCommands
 
 import scala.concurrent.Future
@@ -43,5 +44,8 @@ class DeploymentManagerStub extends BaseDeploymentManager with StubbingCommands 
   )(implicit freshnessPolicy: DataFreshnessPolicy): Future[WithDataFreshnessStatus[List[StatusDetails]]] = {
     Future.successful(WithDataFreshnessStatus.fresh(jobStatus.toList))
   }
+
+  override def getDeploymentStatusesToUpdate: Future[Map[newdeployment.DeploymentId, DeploymentStatus]] =
+    Future.successful(Map.empty)
 
 }