+++ engine/lite/embeddedDeploymentManager/src/main/scala/pl/touk/nussknacker/engine/embedded/streaming/StreamingDeploymentStrategy.scala
@@ -1,8 +1,8 @@
 package pl.touk.nussknacker.engine.embedded.streaming
 
 import com.typesafe.scalalogging.LazyLogging
-import pl.touk.nussknacker.engine.api.deployment.StateStatus
-import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
+import pl.touk.nussknacker.engine.api.deployment.{DeploymentStatus, StateStatus}
+import pl.touk.nussknacker.engine.api.deployment.simple.{SimpleDeploymentStatus, SimpleStateStatus}
 import pl.touk.nussknacker.engine.api.{JobData, LiteStreamMetaData, ProcessVersion}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.embedded.{Deployment, DeploymentStrategy}
@@ -56,10 +56,10 @@ class StreamingDeploymentStrategy extends DeploymentStrategy with LazyLogging {
 
   class StreamingDeployment(interpreter: KafkaTransactionalScenarioInterpreter) extends Deployment {
 
-    override def status(): StateStatus = interpreter.status() match {
-      case TaskStatus.Running      => SimpleStateStatus.Running
-      case TaskStatus.DuringDeploy => SimpleStateStatus.DuringDeploy
-      case TaskStatus.Restarting   => SimpleStateStatus.Restarting
+    override def status(): DeploymentStatus = interpreter.status() match {
+      case TaskStatus.Running      => SimpleDeploymentStatus.Running
+      case TaskStatus.DuringDeploy => SimpleDeploymentStatus.DuringDeploy
+      case TaskStatus.Restarting   => SimpleDeploymentStatus.Restarting
       case other                   => throw new IllegalStateException(s"Not supporter task status: $other")
     }